Module Webdriver_cohttp_lwt_unix.Capabilities

The requested capabilities when creating a new session.

type t = json

A json describing the required capabilities for the session.

val firefox : t

The default firefox configuration:

{ "capabilities":
    { "firstMatch": [{ "browser_name": "firefox" } ] } }

val firefox_headless : t

Same as firefox, but runs in the background without a graphical window:

{ "capabilities":
    { "firstMatch":
        [{ "browser_name": "firefox",
           "moz:firefoxOptions": { "args": ["-headless"] }
         }
        ]
    }
}